0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int countnod;
+	long unsigned int countnod;
 	while (h != NULL)
 	{
 		if (h->str == NULL)
